Move groupByCities to utils
